requirements.txt
numpy==2.2.6
opencv-python==4.11.0.86
ultralytics==8.3.144
psutil==7.0.0
pyzmq==26.4.0
torch==2.7.0
pillow==11.2.1
simpleaudio==1.0.4